backend/requirements.txt
SQLAlchemy==2.0.25
RPi.GPIO==0.7.1
loguru==0.7.2
paho-mqtt==1.6.1
Flask==2.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==2.2.2
smbus2==0.4.3
psycopg2==2.9.9
sqlmodel==0.0.14
line-bot-sdk==3.9.0